kcl-event-grpc-server/requirements.txt
grpcio
grpcio-tools
protobuf
aiokafka